Remove redundant variables for current window
globals defines $WID, $CUR and $PFW although all of these are the same.
Replace all of their instances by $CUR.

This should also fix a bug that prevents snapw from being used with a
windowid: At its start CUR was set to the windowid if one is given, but
later PFW was used.

Moreover, source globals in pulsew. Otherwise it won't work.