feat: implement tilew
tilew is an adaptation of tile.sh in wmutils/contrib.
It takes a wid (default: focused window) and teleports the window to the
left of the screen, using the maximal possible height. The window's
width is specified using the -w option (default: half the display width,
taking into account the middle gap). All other visible windows are
stacked on the right.