feat: add o- and n-options to tilew
o for orientation, n for number of windows in the master-area.
Instead of a w(idth)-option, there is a s(ize)-option now that
determines the size of the master-area. If the orientation is left or
right, this size is the width, but for an orientation that is up or
down, it is the height.