refactor: move window position variables to disp
LEFT, TOP, WHOLEW and WHOLEH are moved to disp such that they can be
used for window positioning that takes into account outer panels and
gaps. They should always be used for such tasks (not SX, SH etc.) since
global changes concerning panels and gaps are then easier to implement.
The other variables in snapw are also defined in terms of them now.