refactor: use single variable for tmp-directory
Attention, config-change!
The temporary files moved to the following places:
  $GRPROOT -> $WMRC_TEMPDIR/groups
  $GRPSTATROOT -> $WMRC_TEMPDIR/grpstat
  $EXPANDDIR -> $WMRC_TEMPDIR/expand
  $FS_FILE -> $WMRC_TEMPDIR/fullscreened

I chose "$XDG_CACHE_HOME"/wmrc as the default for $WMRC_TEMPDIR since it
is user-specific and /tmp should normally only be used for files created
with mkstemp and mkdtemp. The disadvantage is that there is no automatic
cleanup for $XDG_CACHE_HOME.